Update calculator.py
Great work on this. I'm just adding this pull request because I think you might appreciate the style. I offer a couple of suggestions:

```
for i in range(1, len(args)):
            result = result - int(args[i])
```
becomes:
```
for num in args[1:]:
            result *= int(num)
```

In Python, where possible iterate through an array instead of iterating through the indexes of an array. This makes the code prettier, but it also makes the style of your code a little more general: you can iterate through lots of things that don't have numerical indexes, like sets and dictionaries.

```
result = 1
    try:
        for num in args:
            result *= int(num)
```
This is just a math trick that lets you iterate through the entire array in one go, instead of peeling the first number off first. In this case, we start off with `result = 1` because 1 is the "multiplicative identity" or the starting point for any multiplication. You would use 1 as the starting point for multiplication or division, but 0 as the starting point for addition or subtraction.

No changes needed, this pull request is just a way of showing you an alternate way to accomplish what you've done.